feat: usage of default folder code-Rmd instead of codeRmd
Change is in accordance with a naming convention where hyphen is used instead of space. It's more clear e.g. in situations where the capital letter doesn't mean start of a new word, e.g. "R" in codeRmd means start of a new word but "T" in word eToro doesn't mean start of a new word since eToro is one word with one capital letter not as the 1st letter.