feat(workflowrsubdirs): Update the way of using stop() in the script
Commit 4b76068 should also solve the problem when
'Called from: ' appears and debug mode didn't end
when calling stop(). Solution in that commit didn't help.
The problem was in a way of installing the package when
1stly Build Source Package is needed and then installing
the package from built .tar.gz file as it's explained at
https://stackoverflow.com/questions/65004638/stop-in-package-fungction-doesnt-end-debug-mode

parameter call. = F is still used for stop() because
I decided that a user doesn't need to know a name of
function where the error occurs.

There's still one situation (see initial_checks() lines 90-92
and render_html() lines 47-51) where there isn't an error.