feat(workflowrsubdirs): Updates needed for working with figures
This solution is created to avoid situations to not display figures
when functions like graphics::hist() are used for .Rmd files
not saved directly in directory "analysis" (or in case of workflowr subdirs
package not directly in directory "code-rmd") but in subdirectories
like "code-rmd\subdir" (or more subdirectories). If .Rmd files are saved in
any directory within a workflowr working directory then these changes
aren't needed (but in the future I plan to force to be used only "code-rmd").

More relevant information from
https://jdblischak.github.io/workflowr/articles/wflow-04-how-it-works.html:
When functions like graphics::hist() are used then figures are created and
because workflowr requires the figures to be saved to a specific location
in order to function properly, it will override any custom setting of the knitr
option fig.path (which controls where figure files are saved) and
insert a warning into the HTML file to alert the user that their value
for fig.path was ignored.